created the instant class

+++ src/main/java/com/sacco/saccoapp/dividend/DividendRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 @Repository
-public interface DividendRepository extends JpaRepository<Dividend,Integer> {
+public interface DividendRepository extends JpaRepository<Dividend,Long> {
     @Query("SELECT s from Dividend s where s.id =?1")
     Optional<Dividend> findById(Long id);
 }
